openad-model-inference/gt4sd_regression/pyproject.toml
[tool.poetry]
name = "gt4sd_inference_regression"
version = "0.1.0"
license="MIT"
description = "gt4sd Inference for regression"
authors = ["Daniel Choi <choi@ibm.com>", "Phil Downey <Phil.Downey1@ibm.com>", "Brian Duenas <Brian.Duenas@ibm.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.9,<3.11"
gt4sd-trainer-hf-pl = "^1.0.0"
#terminator = {path = "../../regression-transformer", optional = false}
terminator = {git =  "https://github.com/PhillipDowney/regression-transformer", rev = "gt4sd", optional = false}
pytoda = {git = "https://github.com/PhillipDowney/paccmann_datasets", rev = "1.1.3", optional = false}
# gt4sd_common = {path =  "../gt4sd_common/"}
gt4sd_common = { git = "https://github.com/acceleratedscience/Open-AD-Model-Service", subdirectory = "openad-model-inference/gt4sd_common", branch = "excl_moler" }

sentencepiece = "0.2.0"
#psutil = "^5.9.8"
psutil = "^5.9.8"
[tool.poetry.extras]


[tool.poetry.group.dev.dependencies]
pyright = "^1.1.352"
python-lsp-server = {extras = ["all"], version = "^1.10.0"}
mols2grid = { version = "^2.0.0", optional = true}
torch = [
  {url = "https://download.pytorch.org/whl/cu118/torch-2.2.1%2Bcu118-cp310-cp310-linux_x86_64.whl" , platform = "linux"},
  {url = "https://download.pytorch.org/whl/cpu/torch-2.2.1%2Bcpu-cp310-cp310-win_amd64.whl" , platform = "win32" },
  {url = "https://download.pytorch.org/whl/cpu/torch-2.2.1-cp310-none-macosx_10_9_x86_64.whl" , platform = "darwin" }
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
